+++ hss_okta_flutter/example/lib/web/web_profile_screen.dart
@@ -1,11 +1,32 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:jwt_decoder/jwt_decoder.dart';
+
+import '../provider/plugin_provider.dart';
 
 class WebProfileScreen extends StatelessWidget {
-  const WebProfileScreen({super.key, required this.token});
+  const WebProfileScreen({
+    required this.accessToken,
+    super.key,
+    required this.token,
+  });
   final String token;
+  final String accessToken;
   @override
   Widget build(BuildContext context) {
-    return Text(JwtDecoder.decode(token).toString());
+    return Column(
+      children: [
+        OutlinedButton(
+            onPressed: () async {
+              final provider = PluginProvider.of(context);
+              if (kIsWeb) {
+                final isSignedOut = await provider.pluginWeb.signOut();
+                // if (isSignedOut) {
+                //   Navigator.of(context).pushReplacementNamed('/');
+                // }
+              }
+            },
+            child: const Text("Sign out"))
+      ],
+    );
   }
 }