+++ hss_okta_flutter/lib/hss_okta_flutter.dart
@@ -1,12 +1,13 @@
-export 'src/widgets/widgets.dart';
-export 'src/hss_okta_flutter.dart';
 export 'src/enums/sign_in_factors.dart';
 export 'src/generated/hss_okta_flutter.g.dart';
-
+export 'src/hss_okta_flutter.dart';
 export 'src/web/hss_okta_flutter_web.dart'
     if (dart.library.io) 'src/web/stubs/hss_okta_flutter_web_stub.dart'
     if (dart.library.html) 'src/web/hss_okta_flutter_web.dart';
-
+export 'src/web/js/hss_okta_authn_js.dart'
+    if (dart.library.io) 'src/web/stubs/hss_okta_authn_stub.dart'
+    if (dart.library.html) 'src/web/js/hss_okta_authn_js.dart';
 export 'src/web/js/hss_okta_js_external.dart'
     if (dart.library.io) 'src/web/stubs/hss_okta_js_mobile_stub.dart'
     if (dart.library.html) 'src/web/js/hss_okta_js_external.dart';
+export 'src/widgets/widgets.dart';